refactor: MVVM패턴 폴더 구조 정리
