+++ android/app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
     <uses-permission-sdk-23 android:name="android.permission.INTERNET"/>
 
     <application
-        android:name=".MainApplication"
+        android:name=".common.MainApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -14,7 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Airbnb">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
