--- android/app/src/main/java/com/example/airbnb/DataSource.kt
@@ -1,8 +1,0 @@
-package com.example.airbnb
-
-interface DataSource {
-
-    suspend fun loadHeroImage(): HeroImage?
-
-    suspend fun loadCloseTravel(): Travel?
-}
