--- android/app/src/main/java/com/example/airbnb/SearchCloseTravel.kt
@@ -1,8 +1,0 @@
-package com.example.airbnb
-
-data class SearchCloseTravel(
-    val id: Int?,
-    val imageUrl: String?,
-    val title: String?,
-    val contents: String?
-)
