+++ android/app/src/main/java/com/example/airbnb/data/repository/Repository.kt
@@ -1,5 +1,7 @@
-package com.example.airbnb
+package com.example.airbnb.data.repository
 
+import com.example.airbnb.network.dto.HeroImage
+import com.example.airbnb.network.dto.Travel
 import javax.inject.Inject
 
 class Repository @Inject constructor(private val dataSource: DataSource){
