+++ android/app/src/main/java/com/example/airbnb/ui/common/ImageBindingAdapter.kt
