+++ android/app/src/main/java/com/example/airbnb/ui/placesearch/PlaceSearchViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.airbnb
+package com.example.airbnb.ui.placesearch
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
