+++ android/app/src/main/java/com/example/airbnb/ui/wishlist/WishListFragment.kt
@@ -1,10 +1,11 @@
-package com.example.airbnb
+package com.example.airbnb.ui.wishlist
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.example.airbnb.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
