+++ android/app/src/main/res/layout/fragment_my_reservation.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyReservationFragment"
+    tools:context=".ui.myreservation.MyReservationFragment"
     >
 
     <!-- TODO: Update blank fragment layout -->
