+++ android/app/src/main/res/layout/fragment_place_search.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".PlaceSearchFragment">
+        tools:context=".ui.placesearch.PlaceSearchFragment">
 
         <View
             android:id="@+id/view_place_search"
