+++ android/app/src/main/res/layout/item_close_travel.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="travels"
-            type="com.example.airbnb.CloseTravelContents" />
+            type="com.example.airbnb.network.dto.CloseTravelContents" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
